requirements.txt
streamlit
anthropic
langchain
langchain-community
faiss-cpu
pypdf
sentence-transformers
python-dotenv